Bugfixing.
Fixed bug with being Surprised (being able to block when surprised was somewhat of an unintentional-intentional bug).
Player needs to be created before the difficulty (due to us seetting playerHP depending on difficulty)
Fixed a bug with not resetting current-floor-index.
Fixed bosses so they are properly being created (Over-engineered for a future scenario I didn't actually create).